!Converter Fixes (drupal57_to_smf.sql)
git-svn-id: https://devel.simplemachines.org/code/smf/trunk@7869 a96bbafc-d752-43aa-a6db-eed64a5f2cf9

+++ other/converters/drupal57_to_smf.sql
@@ -1,6 +1,6 @@
 ---~ name: "Drupal 5.7 "
 /******************************************************************************/
----~ version: "SMF 2.0"
+---~ version: "SMF 1.1"
 ---~ settings: "/drupal_migration.php"
 ---~ from_prefix: "`" . $drupal_database . "`.$drupal_prefix"
 ---~ table_test: "{$from_prefix}users"
@@ -11,6 +11,15 @@
 TRUNCATE {$to_prefix}members;
 TRUNCATE {$to_prefix}attachments;
 
+DROP TABLE IF EXISTS {$to_prefix}tmp_messages;
+
+CREATE TABLE IF NOT EXISTS {$to_prefix}tmp_messages (
+	old_id_msg int(8) NOT NULL default '0',
+	old_id_topic int(8) NOT NULL default '0',
+	type int(2) NOT NULL default '0',
+	date bigint(15) NOT NULL default '0'
+);
+
 ---* {$to_prefix}members
 SELECT
 	uid AS id_member, SUBSTRING(name, 1, 255) AS member_name,
@@ -18,16 +27,15 @@ SELECT
 	pass AS passwd, '' AS lngfile, '' AS buddy_list, 
 	'' AS pm_ignore_list, '' AS message_labels, '' AS personal_text,
 	'' AS website_title, '' AS website_url, '' AS location, '' AS icq, '' AS aim,
-	'' AS msn, '' AS usertitle, '' AS member_ip, '' AS member_ip2, '' AS secret_question,
-	'' AS additional_groups,
-	access AS last_login, 	
-	IF(uid = '1' , 1, 0) AS id_group,	
-		created AS date_registered, '' AS avatar,
+	'' AS msn, '' AS usertitle, '' AS member_ip, '' AS member_ip2, 
+	'' AS secret_question, '' AS additional_groups, access AS last_login, 	
+	IF(uid = '1' , 1, 0) AS id_group, created AS date_registered, '' AS avatar,
 	SUBSTRING(signature, 1, 65534) AS signature
-FROM {$from_prefix}users where uid > 0;
+FROM {$from_prefix}users WHERE uid > 0;
 ---*
 
 ---{
+/*fixing  id_group for Admins */
 while (true)
 {
 	pastTime($substep);
@@ -39,22 +47,20 @@ while (true)
 			INNER JOIN {$from_prefix}permission AS p ON (r.rid=p.rid)
 		WHERE  p.perm LIKE '%administer forums%'
 		LIMIT $_REQUEST[start], 250");
+	
 	while ($row = convert_fetch_assoc($result))
-	{
 		convert_query("
 			UPDATE {$to_prefix}members
-			SET id_group= 1
+			SET id_group = 1
 			WHERE id_member = $row[id_member]
 			LIMIT 1");
-	}
-
+	
 	$_REQUEST['start'] += 250;
 	if (convert_num_rows($result) < 250)
 		break;
 
 	convert_free_result($result);
 }
-
 $_REQUEST['start'] = 0;
 ---}
 
@@ -64,16 +70,12 @@ $_REQUEST['start'] = 0;
 TRUNCATE {$to_prefix}categories;
 
 ---{
-$no_add = true;
-$keys = array('id_cat', 'name', 'cat_order');
-
 $request = convert_query("
 	SELECT value
 	FROM {$from_prefix}variable
 	WHERE name ='forum_containers'
 	LIMIT 1");
 list($containers) = convert_fetch_row($request);
-convert_free_result($request);
 $cont = unserialize($containers);
 
 foreach ($cont as $categories)
@@ -83,19 +85,17 @@ foreach ($cont as $categories)
 		FROM {$from_prefix}term_data
 		WHERE tid = $categories
 		LIMIT 1");
+		
 	list($name) = convert_fetch_row($request);
-	convert_free_result($request);
-
+	
 	convert_query("
 		INSERT IGNORE INTO {$to_prefix}categories
 			(id_cat, name, catorder)
 		VALUES ('$categories' , '$name' ,0)");
 }
----}
+convert_free_result($request);
 
----{
-$no_add = true;
-$keys = array('id_cat', 'cat_order');
+/* the cat_order is wrong and must be fixed..*/
 $neworder = -1;
 
 $request = convert_query("
@@ -103,13 +103,13 @@ $request = convert_query("
 	FROM {$to_prefix}categories AS c
 		INNER JOIN {$from_prefix}term_data AS t ON (c.id_cat = t.tid)
 	ORDER BY t.weight ASC, c.name ASC");
-
-while ($row= convert_fetch_assoc($request))
+	
+while ($row = convert_fetch_assoc($request))
 	convert_query("
 		UPDATE {$to_prefix}categories
 		SET cat_order = " . ++$neworder . "
 		WHERE id_cat = $row[id_cat]");
-
+	
 convert_free_result($request);
 ---}
 
@@ -118,62 +118,47 @@ convert_free_result($request);
 /******************************************************************************/
 TRUNCATE {$to_prefix}boards;
 DELETE FROM {$to_prefix}board_permissions
-WHERE id_profile > 4;
+WHERE id_board != 0;
 
 ---* {$to_prefix}boards
----{
-/*!!! CONVERT THIS FROM MYSQL SPECIFIC QUERY!!! */
-/* WHY ARE WE REALLY DOING IT THIS WAY? LET CONVERT.PHP HANDLE THIS */
-
-$no_add = true;
-$keys = array('id_board', 'name', 'description', 'id_parent', 'id_cat');
-$request = convert_query("
-	INSERT IGNORE INTO {$to_prefix}boards
-		(id_board, name, description, id_parent, id_cat)
-		VALUES ('$row[id_board]' , '$row[name]' , '$row[description]' , '$row[id_parent]' , '$row[id_cat]')");
-convert_free_result($request);
----}
-
 SELECT t.tid AS id_board, SUBSTRING(t.name, 1, 255) AS name, SUBSTRING(t.description, 1, 255) AS description, 
-	IF(h.parent=c.id_cat, 0, h.parent) AS id_parent, 
+	IF(h.parent = c.id_cat, 0, h.parent) AS id_parent, 
 	IF(c.id_cat IS NULL , 0, c.id_cat) AS id_cat
 FROM {$from_prefix}term_data AS t
-	INNER JOIN {$from_prefix}term_hierarchy AS h ON (t.tid = h.tid)
+	LEFT JOIN {$from_prefix}term_hierarchy AS h ON (t.tid = h.tid)
 	LEFT JOIN {$to_prefix}categories AS c ON (h.parent = c.id_cat);
 ---*
 
 ---{
-$no_add = true;
-$keys = array('id_cat', 'name', 'cat_order');
-
+/* Some boards are categories and should not be here ...*/
 $request = convert_query("
 	SELECT value
 	FROM {$from_prefix}variable
 	WHERE name ='forum_containers'
 	LIMIT 1");
 list($containers) = convert_fetch_row($request);
-convert_free_result($request);
+
 $cont = unserialize($containers);
 
 foreach ($cont as $categories)
 {
+
 	$request = convert_query("
 		SELECT name
 		FROM {$from_prefix}term_data
 		WHERE tid = $categories
 		LIMIT 1");
 	list($name) = convert_fetch_row($request);
-	convert_free_result($request);
 
-	convert_query("
-		DELETE FROM {$to_prefix}boards
-		WHERE id_board = $categories ");
+/* now we delete them...*/	
+	if(isset($name))
+		convert_query("
+			DELETE FROM {$to_prefix}boards
+			WHERE id_board = $categories ");
 }
----}
+convert_free_result($request);
 
----{
-$no_add = true;
-$keys = array('id_board', 'board_order');
+/* the board_order is wrong and must be fixed..*/
 $neworder = -1;
 
 $request = convert_query("
@@ -187,8 +172,29 @@ while ($row= convert_fetch_assoc($request))
 		UPDATE {$to_prefix}boards
 		SET board_order = " . ++$neworder . "
 		WHERE id_board = $row[id_board]");
+		
 convert_free_result($request);
 ---}
+
+/******************************************************************************/
+--- preparing messages ...
+/******************************************************************************/
+---* {$to_prefix}tmp_messages
+
+SELECT
+	nid AS old_id_msg, nid AS old_id_topic, '1' AS type, created AS date
+FROM {$from_prefix}node;
+---*
+
+/******************************************************************************/
+--- preparing messages  (part 2)...
+/******************************************************************************/
+---* {$to_prefix}tmp_messages
+
+SELECT
+	cid AS old_id_msg, nid AS old_id_topic, '2' AS type, timestamp AS date
+FROM {$from_prefix}comments;
+---*
 
 /******************************************************************************/
 --- Converting topics...
@@ -198,37 +204,30 @@ TRUNCATE {$to_prefix}log_topics;
 TRUNCATE {$to_prefix}log_boards;
 TRUNCATE {$to_prefix}log_mark_read;
 
+ALTER TABLE {$to_prefix}tmp_messages ORDER BY date;
+ALTER TABLE {$to_prefix}tmp_messages 
+	ADD id_msg INT( 12 ) NOT NULL AUTO_INCREMENT PRIMARY KEY FIRST;
+
 ---* {$to_prefix}topics
----{
-/*!!! CONVERT THIS FROM MYSQL SPECIFIC QUERY!!! */
-/* WHY ARE WE REALLY DOING IT THIS WAY? LET CONVERT.PHP HANDLE THIS */
-
-$no_add = true;
-$keys = array('id_topic', 'id_board', 'is_sticky', 'id_first_msg', 'id_last_msg', 'id_member_started', 'id_member_updated');
-
-$request = convert_query("
-	INSERT IGNORE INTO {$to_prefix}topics
-		(id_topic, id_board, is_sticky, id_first_msg, id_last_msg, id_member_started, id_member_updated)
-		VALUES ('$row[id_topic]' , '$row[id_board]' , '$row[is_sticky]' , '$row[id_first_msg]' , '$row[id_last_msg]' , '$row[id_member_started]' , '$row[id_member_updated]')");
-convert_free_result($request);	
----}
 
 SELECT
 	t.nid AS id_topic, f.tid AS id_board, t.sticky AS is_sticky,
-	t.nid AS id_first_msg, 
-	IF (MAX(c.nid)>0, MAX(c.nid), t.nid) AS id_last_msg,
-	t.uid AS id_member_started,
-	IF (MAX(c.uid)>0, MAX(c.uid), t.uid) AS id_member_updated 
+	MIN(id.id_msg) AS id_first_msg, MAX(id.id_msg) AS id_last_msg,
+	t.uid AS id_member_started,	IF (MAX(c.uid)>0, MAX(c.uid), t.uid) AS id_member_updated 
 FROM {$from_prefix}node AS t
+	INNER JOIN {$to_prefix}tmp_messages AS id ON (t.nid = id.old_id_topic)
 	INNER JOIN {$from_prefix}forum AS f ON (f.nid = t.nid)
 	LEFT JOIN {$from_prefix}comments AS c ON (c.nid = t.nid)
-WHERE t.type='forum'
-GROUP BY t.nid;
+WHERE t.type = 'forum'
+GROUP BY t.nid
+HAVING id_first_msg != 0
+	AND id_last_msg != 0;
 ---*
 
 /******************************************************************************/
 --- Converting posts (this may take some time)...
 /******************************************************************************/
+
 TRUNCATE {$to_prefix}messages;
 TRUNCATE {$to_prefix}attachments;
 
@@ -294,48 +293,43 @@ $row['body'] = preg_replace(
 ---}
 
 SELECT
-	p.nid AS id_msg, p.nid AS id_topic, f.tid AS id_board, p.uid AS id_member,
-	p.timestamp AS poster_time,
-	t.title AS subject,
-	p.body AS body,
-	u.name AS poster_name, '' as poster_ip, '' as modified_name,
-	u.mail AS poster_email,
-	'xx' AS icon
-FROM {$from_prefix}node_revisions AS p 
-	INNER JOIN {$from_prefix}node AS t ON (p.nid = t.nid)
+	id.id_msg AS id_msg, p.nid AS id_topic, f.tid AS id_board, p.uid AS id_member,
+	p.created AS poster_time, p.title AS subject, t.body AS body, u.name AS poster_name,
+	'' as poster_ip, '' as modified_name,	u.mail AS poster_email, 'xx' AS icon
+FROM {$to_prefix}tmp_messages AS id
+	INNER JOIN {$from_prefix}node AS p ON (p.nid = id.old_id_msg)
+	INNER JOIN {$from_prefix}node_revisions AS t ON (p.nid = t.nid)
 	INNER JOIN {$from_prefix}forum AS f ON (t.nid = f.nid)
 	INNER JOIN {$from_prefix}users AS u ON (u.uid = t.uid)
-WHERE t.type='forum';
-
+WHERE id.type = '1' AND p.type = 'forum'
+GROUP BY id.id_msg;
 ---*
 /******************************************************************************/
---- Converting posts - Part 2(this may take some time)...
+--- Converting posts - Part 2 (this may take some time)...
 /******************************************************************************/
 
 ---* {$to_prefix}messages 200
 SELECT
-	p.nid AS id_topic, f.tid AS id_board, c.uid AS id_member,
-	c.timestamp AS poster_time,
-	c.subject AS subject,
-	c.comment AS body, '' AS poster_ip, '' AS modified_name,
-	c.name AS poster_name,
-	u.mail AS poster_email,
+	id.id_msg AS id_msg, p.nid AS id_topic, f.tid AS id_board, c.uid AS id_member, 
+	c.timestamp AS poster_time, c.subject AS subject, c.comment AS body, 
+	'' AS poster_ip, '' AS modified_name,	c.name AS poster_name, u.mail AS poster_email,
 	'xx' AS icon
-FROM {$from_prefix}comments AS c
+FROM {$to_prefix}tmp_messages AS id
+	INNER JOIN {$from_prefix}comments AS c ON (c.cid = id.old_id_msg)
 	INNER JOIN {$from_prefix}node AS t ON (c.nid = t.nid)
 	INNER JOIN {$from_prefix}node_revisions AS p ON (p.nid = t.nid)
 	INNER JOIN {$from_prefix}forum AS f ON (t.nid = f.nid)
-	INNER JOIN {$from_prefix}users AS u ON (u.uid=c.uid)
-WHERE t.type='forum'
+	INNER JOIN {$from_prefix}users AS u ON (u.uid = c.uid)
+WHERE t.type = 'forum' AND id.type = '2'
+GROUP BY id.id_msg
 ORDER BY c.cid;
 ---*
 
----# Get all the members posts...
+---# Get all the members posts and fix them ...
 ---{
 while (true)
 {
 	pastTime($substep);
-
 	$result = convert_query("
 		SELECT m.id_member, COUNT(m.id_msg) AS posts
 		FROM {$to_prefix}messages AS m
@@ -343,7 +337,7 @@ while (true)
 		WHERE  b.count_posts = 0
 		GROUP BY m.id_member
 		LIMIT $_REQUEST[start], 250");
-	while ($row = convert_fetch_assoc($result))
+	while ($row = mysql_fetch_assoc($result))
 	{
 		$row['posts'] = (int) $row['posts'];
 
@@ -383,15 +377,17 @@ if (file_exists($oldfile) && strlen($newfilename) <= 255 && copy($_POST['path_fr
 	$rows[] = "$id_attach, " . filesize($attachmentUploadDir . '/' . $newfilename) . ", '" . addslashes(basename($row['filename'])) . "', $row[id_msg], 0";
 	$id_attach++;
 }
-
 ---}
-SELECT nid AS id_msg, filename, filepath
-FROM {$from_prefix}files;
+SELECT 
+	id.id_msg AS id_msg, f.filename, f.filepath 
+FROM {$from_prefix}files AS f
+	INNER JOIN {$to_prefix}tmp_messages AS id ON (f.nid = id.old_id_msg);
 ---*
 
 /******************************************************************************/
 --- Converting avatars...
 /******************************************************************************/
+DROP TABLE IF EXISTS {$to_prefix}tmp_messages;
 
 ---* {$to_prefix}attachments
 
@@ -409,7 +405,9 @@ if (strlen($newfilename) <= 255 && copy($_POST['path_from'] . '/' . $filepath , 
 }
 ---}
 
-SELECT uid AS id_member, picture 
+SELECT 
+	uid AS id_member, picture 
 FROM {$from_prefix}users
 WHERE picture != '';
+
 ---*
